Move environment config from template to maji source
